Deploy cadastre + api tracking + mock for search test

+++ app/batid/services/source.py
@@ -2,6 +2,7 @@ import csv
 import json
 import os
 import tarfile
+import gzip
 
 import nanoid
 import py7zr
@@ -13,7 +14,7 @@ class Source:
 
     # Must be prefixed with a dot
 
-    archive_exts = [".7z", ".tar.gz"]
+    archive_exts = [".7z", ".tar.gz", ".gz"]
 
     def __init__(self, name, custom_ref=None):
         self.name = name
@@ -28,6 +29,11 @@ class Source:
 
     def default_ref(self) -> dict:
         return {
+            "plot": {
+                "folder": "plots",
+                "url": "https://cadastre.data.gouv.fr/data/etalab-cadastre/2023-07-01/geojson/departements/{{dpt}}/cadastre-{{dpt}}-parcelles.json.gz",
+                "filename": "cadastre-{{dpt}}-parcelles.json",
+            },
             "bdnb_7_buffer": {
                 "folder": "bdnb_7",
                 "filename": "buffer.csv",
@@ -72,7 +78,10 @@ class Source:
         if "dl_filename" in self.ref:
             return self.ref["dl_filename"]
 
-        return os.path.basename(self.url)
+        if "url" in self.ref:
+            return os.path.basename(self.url)
+
+        return None
 
     @property
     def path(self) -> str:
@@ -112,8 +121,11 @@ class Source:
 
     @property
     def is_archive(self):
+        filename = self.dl_filename if self.dl_filename is not None else self.filename
+
         for ext in self.archive_exts:
-            if self.dl_filename.endswith(ext):
+            # First we check the downloaded file
+            if filename.endswith(ext):
                 return True
 
         return False
@@ -137,11 +149,15 @@ class Source:
 
         if self.dl_filename.endswith(".7z"):
             self.uncompress_7z()
+            return
 
         if self.dl_filename.endswith(".tar.gz"):
             self.uncompress_tar_gz()
+            return
 
-        self.remove_archive()
+        if self.dl_filename.endswith(".gz"):
+            self.uncompress_gz()
+            return
 
     def uncompress_7z(self):
         with py7zr.SevenZipFile(self.dl_path, "r") as archive:
@@ -151,7 +167,15 @@ class Source:
         with tarfile.open(self.dl_path, "r:gz") as tar:
             tar.extractall(self.uncompress_abs_dir)
 
+    def uncompress_gz(self):
+        with gzip.open(self.dl_path, "rb") as f_in:
+            with open(self.path, "wb") as f_out:
+                f_out.write(f_in.read())
+
     def remove_archive(self):
+        if not self.is_archive:
+            return
+
         os.remove(self.dl_path)
 
     def find(self, filename):
